// frontend/app/interfaces/order.interface.ts
export enum OrderTypeTitle {
    OPEN = 'Очик буюртмалар',
    TOMORROW = 'Эртага бажарилиши лозим буюртмалар',
    DELETED = 'Учирилган буюртмалар',
    COMPLETED = 'Бажарилган буюртмалар',
    EXPIRED = 'Муддати утган буюртмалар'
}

export enum OrderStatus {
    OPEN = 'OPEN',
    DELETED = 'DELETED',
    COMPLETED = 'COMPLETED',
    EXPIRED = 'EXPIRED'
}